chore(store): rename folder "create-logux" to "store"
