+++ store/index.d.ts
